Continued reorganizing exmaple and benchmark files
